Refactor: move std lib configuration to ArendTestBase
