+++ src/test/resources/org/arend/arend-lib/src/Function.ard
